requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
pydantic==1.9.0
opencv-python>=4.1.2
python-multipart>=0.0.5

# YOLOv5 requirements
# https://github.com/ultralytics/yolov5/blob/master/requirements.txt

matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.64.0

pandas>=1.1.4
seaborn>=0.11.0

onnx>=1.12.0
onnxruntime>=1.12.1
psutil>=5.9.2
ipython>=8.5.0